+++ doc/changelog.md
@@ -1,3 +1,9 @@
+## 0.9.3
+#### documentation Changes
+ - Wrote documentation for elements in [staging.js](fallen/staging.js.md)
+#### Fallen Changes
+ - Moved all elements in staging to input
+
 ## 0.9.2
 #### Fallen Changes
  - Added more `tester.js` checks to staging